requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==3.2.19
django-phonenumber-field==6.4.0
djangorestframework==3.13.1
gunicorn==21.2.0
packaging==23.2
phonenumberslite==8.13.23
Pillow==9.1.1
psycopg2-binary==2.9.3
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3